+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-06-09
+### Added
+- Added support for enriching IP addresses with the ipapi.is API.
+### Changed
+- The default output format for Team Cymru data was set to CSV.
+### Removed
+- Removed support for writing the output to a file. 
+
 ## [1.0.2] - 2023-02-15
 ### Fixed
 - Make input handling more robust